// WYNKWeb/src/app/Models/configure.model.ts
      
export class Configure {

  ID: number;
  CMPID: number;  
  RRDescription: string;
  RRAdvDays: number;
  FrequencyperDay: number;
  HostEmailID: string;
  HostPassword: string;
  Phonenumber: string;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  SendSMS: number;
  SendEmail: number;
}